+++ spec/github/client/repos/keys/create_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe Github::Client::Repos::Keys, '#create' do
+RSpec.describe Github::Client::Repos::Keys, '#create' do
   let(:user) { 'peter-murach' }
   let(:repo) { 'github' }
   let(:request_path) { "/repos/#{user}/#{repo}/keys" }
@@ -10,8 +10,8 @@ describe Github::Client::Repos::Keys, '#create' do
 
   before {
     stub_post(request_path).with(inputs).
-      to_return(:body => body, :status => status,
-        :headers => {:content_type => "application/json; charset=utf-8"})
+      to_return(body: body, status: status,
+        headers: {content_type: "application/json; charset=utf-8"})
   }
 
   after { reset_authentication_for(subject) }
@@ -20,26 +20,14 @@ describe Github::Client::Repos::Keys, '#create' do
     let(:body) { fixture("repos/key.json") }
     let(:status) { 201 }
 
-    it "should fail to create resource if 'title' input is missing" do
-      expect {
-        subject.create user, repo, :key => 'ssh-rsa AAA...'
-      }.to raise_error(Github::Error::RequiredParams)
-    end
-
-    it "should fail to create resource if 'key' input is missing" do
-      expect {
-        subject.create user, repo, :title => 'octocat@octomac'
-      }.to raise_error(Github::Error::RequiredParams)
-    end
-
     it "should create the resource" do
-      subject.create user, repo, inputs
-      a_post(request_path).with(inputs).should have_been_made
+      subject.create(user, repo, inputs)
+      expect(a_post(request_path).with(inputs)).to have_been_made
     end
 
     it "should get the key information back" do
-      key = subject.create user, repo, inputs
-      key.title.should == 'octocat@octomac'
+      key = subject.create(user, repo, inputs)
+      expect(key.title).to eq('octocat@octomac')
     end
   end