+++ spec/github/client/repos/keys_spec.rb
@@ -2,8 +2,6 @@
 
 require 'spec_helper'
 
-describe Github::Client::Repos::Keys do
+RSpec.describe Github::Client::Repos::Keys do
   it_should_behave_like 'api interface'
-
-  it { described_class::VALID_KEY_OPTIONS.should_not be_nil }
 end # Github::Repos::Keys